tortoise: fix fetch list

+++ cvsnt/tortoiseCVS/TortoiseCVS/tools/cppunit/src/cppunit/CompilerOutputter.cpp
@@ -139,7 +139,7 @@ CompilerOutputter::processLocationFormatCommand( char command,
 std::string 
 CompilerOutputter::extractBaseName( const std::string &fileName ) const
 {
-  int indexLastDirectorySeparator = fileName.find_last_of( '/' );
+  intptr_t indexLastDirectorySeparator = fileName.find_last_of( '/' );
   
   if ( indexLastDirectorySeparator < 0 )
     indexLastDirectorySeparator = fileName.find_last_of( '\\' );
